// src/app/state/event-type.enum.ts
export enum EventType {

  LOGIN = "LOGIN",
  REGISTER_TALENT = "REGISTER_TALENT",
  REGISTER_CLIENT = "REGISTER_CLIENT",
  RECOVER_PASSWORD = "RECOVER_PASSWORD",
  RESET_PASSWORD = "RESET_PASSWORD",
  CONFIRM_ACCOUNT = "CONFIRM_ACCOUNT",

  GET_JOBS = "GET_JOBS",
  GET_JOB = "GET_JOB",
  GET_JOB_APPLICATIONS = "GET_JOB_APPLICATIONS",
  GET_JOB_SELECTED_APPLICATIONS = "GET_JOB_SELECTED_APPLICATIONS",

  UPDATE_TALENT_PROFILE = "UPDATE_TALENT_PROFILE",

  OPEN_ADD_EXPERIENCE = "OPEN_ADD_EXPERIENCE",
  CLOSE_ADD_EXPERIENCE = "CLOSE_ADD_EXPERIENCE",
  OPEN_EDIT_EXPERIENCE = "OPEN_EDIT_EXPERIENCE",
  CLOSE_EDIT_EXPERIENCE = "CLOSE_EDIT_EXPERIENCE",

  ADD_EXPERIENCE = "ADD_EXPERIENCE",
  UPDATE_EXPERIENCE = "UPDATE_EXPERIENCE",
  DELETE_EXPERIENCE = "DELETE_EXPERIENCE",

  OPEN_ADD_EDUCATION = "OPEN_ADD_EDUCATION",
  CLOSE_ADD_EDUCATION = "CLOSE_ADD_EDUCATION",
  OPEN_EDIT_EDUCATION = "OPEN_EDIT_EDUCATION",
  CLOSE_EDIT_EDUCATION = "CLOSE_EDIT_EDUCATION",

  ADD_EDUCATION = "ADD_EDUCATION ",
  UPDATE_EDUCATION = "UPDATE_EDUCATION",
  DELETE_EDUCATION = "DELETE_EDUCATION",

  OPEN_ADD_PROJECT = "OPEN_ADD_PROJECT",
  CLOSE_ADD_PROJECT = "CLOSE_ADD_PROJECT",
  OPEN_EDIT_PROJECT = "OPEN_EDIT_PROJECT",
  CLOSE_EDIT_PROJECT = "CLOSE_EDIT_PROJECT",

  ADD_PROJECT = "ADD_PROJECT",
  UPDATE_PROJECT = "UPDATE_PROJECT",
  DELETE_PROJECT = "DELETE_PROJECT",

  OPEN_ADD_LANGUAGE = "OPEN_ADD_LANGUAGE",
  CLOSE_ADD_LANGUAGE = "CLOSE_ADD_LANGUAGE",
  OPEN_EDIT_LANGUAGE = "OPEN_EDIT_LANGUAGE",
  CLOSE_EDIT_LANGUAGE = "CLOSE_EDIT_LANGUAGE",

  ADD_LANGUAGE = "ADD_LANGUAGE",
  UPDATE_LANGUAGE = "UPDATE_LANGUAGE",
  DELETE_LANGUAGE = "DELETE_LANGUAGE",

  OPEN_ADD_CERTIFICATION = "OPEN_ADD_CERTIFICATION",
  CLOSE_ADD_CERTIFICATION = "CLOSE_ADD_CERTIFICATION",
  OPEN_EDIT_CERTIFICATION = "OPEN_EDIT_CERTIFICATION",
  CLOSE_EDIT_CERTIFICATION = "CLOSE_EDIT_CERTIFICATION",

  ADD_CERTIFICATION = "ADD_CERTIFICATION",
  UPDATE_CERTIFICATION = "UPDATE_CERTIFICATION",
  DELETE_CERTIFICATION = "DELETE_CERTIFICATION",

  UPDATE_CLIENT_PROFILE = "UPDATE_CLIENT_PROFILE",

  OPEN_ADD_JOB = "OPEN_ADD_JOB",
  CLOSE_ADD_JOB = "CLOSE_ADD_JOB",
  OPEN_EDIT_JOB = "OPEN_EDIT_JOB",
  CLOSE_EDIT_JOB = "CLOSE_EDIT_JOB",

  ADD_JOB = "ADD_JOB",
  UPDATE_JOB = "UPDATE_JOB",
  DELETE_JOB = "DELETE_JOB",

  CHANGE_PAGE = "CHANGE_PAGE",

  UPDATE_IMAGE = "UPDATE_IMAGE",
  ASK_VERIFICATION = "ASK_VERIFICATION",


  OPEN_JOB = "OPEN_JOB",
  CLOSE_JOB = "CLOSE_JOB",
  APPLY_TO_JOB = "APPLY_TO_JOB",
}